Implemented working scroll syncing again
